finished all the levels logic started to work on level design

+++ Ass6/src/game/animation/Animation.java
@@ -12,7 +12,7 @@ public interface Animation {
 
     void drawBackground(DrawSurface canvas);
 
-    default int getFramePerSeconds() {
+    default double getFramePerSeconds() {
         return FRAMES_PER_SECONDS;
     }
 }
